ptrace: replace ptrace kprobes with security_ptrace_access_check
