+++ Makefile
@@ -70,18 +70,16 @@ p_lkrg-objs += src/modules/ksyms/p_resolve_ksym.o \
                src/modules/exploit_detection/syscalls/keyring/p_sys_add_key/p_sys_add_key.o \
                src/modules/exploit_detection/syscalls/keyring/p_sys_request_key/p_sys_request_key.o \
                src/modules/exploit_detection/syscalls/keyring/p_sys_keyctl/p_sys_keyctl.o \
-               src/modules/exploit_detection/syscalls/p_sys_ptrace/p_sys_ptrace.o \
+               src/modules/exploit_detection/syscalls/p_security_ptrace_access/p_security_ptrace_access.o \
                src/modules/exploit_detection/syscalls/compat/p_compat_sys_execve/p_compat_sys_execve.o \
                src/modules/exploit_detection/syscalls/compat/p_compat_sys_execveat/p_compat_sys_execveat.o \
                src/modules/exploit_detection/syscalls/compat/p_compat_sys_keyctl/p_compat_sys_keyctl.o \
-               src/modules/exploit_detection/syscalls/compat/p_compat_sys_ptrace/p_compat_sys_ptrace.o \
                src/modules/exploit_detection/syscalls/compat/p_compat_sys_capset/p_compat_sys_capset.o \
                src/modules/exploit_detection/syscalls/compat/p_compat_sys_add_key/p_compat_sys_add_key.o \
                src/modules/exploit_detection/syscalls/compat/p_compat_sys_request_key/p_compat_sys_request_key.o \
                src/modules/exploit_detection/syscalls/__x32/p_x32_sys_execve/p_x32_sys_execve.o \
                src/modules/exploit_detection/syscalls/__x32/p_x32_sys_execveat/p_x32_sys_execveat.o \
                src/modules/exploit_detection/syscalls/__x32/p_x32_sys_keyctl/p_x32_sys_keyctl.o \
-               src/modules/exploit_detection/syscalls/__x32/p_x32_sys_ptrace/p_x32_sys_ptrace.o \
                src/modules/exploit_detection/syscalls/override/p_override_creds/p_override_creds.o \
                src/modules/exploit_detection/syscalls/override/p_revert_creds/p_revert_creds.o \
                src/modules/exploit_detection/syscalls/override/overlayfs/p_ovl_create_or_link/p_ovl_create_or_link.o \
