+++ src/modules/exploit_detection/p_exploit_detection.c
@@ -85,6 +85,13 @@ static const struct p_functions_hooks {
      NULL,
      1
    },
+   { "security_ptrace_access",
+     p_install_security_ptrace_access_hook,
+     p_uninstall_security_ptrace_access_hook,
+     0,
+     "LKRG won't enforce validation on 'security_ptrace_access'",
+     0
+   },
    { "sys_setuid",
      p_install_sys_setuid_hook,
      p_uninstall_sys_setuid_hook,
@@ -222,13 +229,6 @@ static const struct p_functions_hooks {
      NULL,
      0
    },
-   { "sys_ptrace",
-     p_install_sys_ptrace_hook,
-     p_uninstall_sys_ptrace_hook,
-     0,
-     "LKRG won't enforce validation on 'sys_ptrace'",
-     0
-   },
 #ifdef CONFIG_COMPAT
    { "compat_sys_execve",
      p_install_compat_sys_execve_hook,
@@ -251,13 +251,6 @@ static const struct p_functions_hooks {
      p_uninstall_compat_sys_keyctl_hook,
      1,
      NULL,
-     0
-   },
-   { "compat_sys_ptrace",
-     p_install_compat_sys_ptrace_hook,
-     p_uninstall_compat_sys_ptrace_hook,
-     0,
-     "LKRG won't enforce validation on 'compat_sys_ptrace'",
      0
    },
  #ifdef P_SYSCALL_LAYOUT_4_17
@@ -307,13 +300,6 @@ static const struct p_functions_hooks {
      p_uninstall_x32_sys_keyctl_hook,
      1,
      NULL,
-     0
-   },
-   { "x32_sys_ptrace",
-     p_install_x32_sys_ptrace_hook,
-     p_uninstall_x32_sys_ptrace_hook,
-     0,
-     "LKRG won't enforce validation on 'x32_sys_ptrace'",
      0
    },
  #endif /* P_SYSCALL_LAYOUT_4_17 */
