+++ src/modules/exploit_detection/p_exploit_detection.h
@@ -269,18 +269,16 @@ struct p_ed_global_variables {
 #include "syscalls/keyring/p_sys_add_key/p_sys_add_key.h"
 #include "syscalls/keyring/p_sys_request_key/p_sys_request_key.h"
 #include "syscalls/keyring/p_sys_keyctl/p_sys_keyctl.h"
-#include "syscalls/p_sys_ptrace/p_sys_ptrace.h"
+#include "syscalls/p_security_ptrace_access/p_security_ptrace_access.h"
 #include "syscalls/compat/p_compat_sys_execve/p_compat_sys_execve.h"
 #include "syscalls/compat/p_compat_sys_execveat/p_compat_sys_execveat.h"
 #include "syscalls/compat/p_compat_sys_keyctl/p_compat_sys_keyctl.h"
-#include "syscalls/compat/p_compat_sys_ptrace/p_compat_sys_ptrace.h"
 #include "syscalls/compat/p_compat_sys_capset/p_compat_sys_capset.h"
 #include "syscalls/compat/p_compat_sys_add_key/p_compat_sys_add_key.h"
 #include "syscalls/compat/p_compat_sys_request_key/p_compat_sys_request_key.h"
 #include "syscalls/__x32/p_x32_sys_execve/p_x32_sys_execve.h"
 #include "syscalls/__x32/p_x32_sys_execveat/p_x32_sys_execveat.h"
 #include "syscalls/__x32/p_x32_sys_keyctl/p_x32_sys_keyctl.h"
-#include "syscalls/__x32/p_x32_sys_ptrace/p_x32_sys_ptrace.h"
 /* Override creds */
 #include "syscalls/override/p_override_creds/p_override_creds.h"
 #include "syscalls/override/p_revert_creds/p_revert_creds.h"
