--- src/modules/exploit_detection/syscalls/__x32/p_x32_sys_ptrace/p_x32_sys_ptrace.h
@@ -1,43 +1,0 @@
-/*
- * pi3's Linux kernel Runtime Guard
- *
- * Component:
- *  - Intercept X32 ptrace syscall
- *
- * Notes:
- *  - Enforce Exploit Detection validation
- *
- * Caveats:
- *  - None
- *
- * Timeline:
- *  - Created: 13.VIII.2018
- *
- * Author:
- *  - Adam 'pi3' Zabrocki (http://pi3.com.pl)
- *
- */
-
-#ifdef CONFIG_X86_X32
-
-#ifdef P_SYSCALL_LAYOUT_4_17
-
-#ifndef P_LKRG_EXPLOIT_DETECTION_X32_SYS_PTRACE_H
-#define P_LKRG_EXPLOIT_DETECTION_X32_SYS_PTRACE_H
-
-/* per-instance private data */
-struct p_x32_sys_ptrace_data {
-    ktime_t entry_stamp;
-};
-
-
-int p_x32_sys_ptrace_ret(struct kretprobe_instance *p_ri, struct pt_regs *p_regs);
-int p_x32_sys_ptrace_entry(struct kretprobe_instance *p_ri, struct pt_regs *p_regs);
-int p_install_x32_sys_ptrace_hook(int p_isra);
-void p_uninstall_x32_sys_ptrace_hook(void);
-
-#endif
-
-#endif
-
-#endif
