--- src/modules/exploit_detection/syscalls/compat/p_compat_sys_ptrace/p_compat_sys_ptrace.c
@@ -1,77 +1,0 @@
-/*
- * pi3's Linux kernel Runtime Guard
- *
- * Component:
- *  - Intercept compat_ptrace syscall
- *
- * Notes:
- *  - Enforce Exploit Detection validation
- *
- * Caveats:
- *  - None
- *
- * Timeline:
- *  - Created: 31.I.2018
- *
- * Author:
- *  - Adam 'pi3' Zabrocki (http://pi3.com.pl)
- *
- */
-
-#ifdef CONFIG_COMPAT
-
-#include "../../../../../p_lkrg_main.h"
-
-
-char p_compat_sys_ptrace_kretprobe_state = 0;
-
-static struct kretprobe p_compat_sys_ptrace_kretprobe = {
-    .kp.symbol_name = P_GET_COMPAT_SYSCALL_NAME(ptrace),
-    .handler = p_compat_sys_ptrace_ret,
-    .entry_handler = p_compat_sys_ptrace_entry,
-    .data_size = sizeof(struct p_compat_sys_ptrace_data),
-    /* Probe up to 40 instances concurrently. */
-    .maxactive = 40,
-};
-
-/*
- * x86-64 syscall ABI:
- *  *rax - syscall_number
- *    rdi - 1st argument
- *    rsi - 2nd argument
- *    rdx - 3rd argument
- *    rcx - 4th argument
- *
- *    r8  - 5th one
- *    r9  - 6th one
- */
-
-int p_compat_sys_ptrace_entry(struct kretprobe_instance *p_ri, struct pt_regs *p_regs) {
-
-   struct p_ed_process *p_tmp;
-   unsigned long p_flags;
-
-   p_tasks_read_lock(&p_flags);
-   if ( (p_tmp = p_find_ed_by_pid(task_pid_nr(current))) != NULL) {
-      // This process is on the ED list - validate 'off' flag
-      p_ed_is_off_off_wrap(p_tmp);
-   }
-   p_tasks_read_unlock(&p_flags);
-
-   p_ed_enforce_validation();
-
-   return 0;
-}
-
-
-int p_compat_sys_ptrace_ret(struct kretprobe_instance *p_ri, struct pt_regs *p_regs) {
-
-//   p_ed_enforce_validation();
-
-   return 0;
-}
-
-
-GENERATE_INSTALL_FUNC(compat_sys_ptrace)
-
-#endif
