--- src/modules/exploit_detection/syscalls/compat/p_compat_sys_ptrace/p_compat_sys_ptrace.h
@@ -1,39 +1,0 @@
-/*
- * pi3's Linux kernel Runtime Guard
- *
- * Component:
- *  - Intercept compat_ptrace syscall
- *
- * Notes:
- *  - Enforce Exploit Detection validation
- *
- * Caveats:
- *  - None
- *
- * Timeline:
- *  - Created: 31.I.2018
- *
- * Author:
- *  - Adam 'pi3' Zabrocki (http://pi3.com.pl)
- *
- */
-
-#ifdef CONFIG_COMPAT
-
-#ifndef P_LKRG_EXPLOIT_DETECTION_COMPAT_SYS_PTRACE_H
-#define P_LKRG_EXPLOIT_DETECTION_COMPAT_SYS_PTRACE_H
-
-/* per-instance private data */
-struct p_compat_sys_ptrace_data {
-    ktime_t entry_stamp;
-};
-
-
-int p_compat_sys_ptrace_ret(struct kretprobe_instance *p_ri, struct pt_regs *p_regs);
-int p_compat_sys_ptrace_entry(struct kretprobe_instance *p_ri, struct pt_regs *p_regs);
-int p_install_compat_sys_ptrace_hook(int p_isra);
-void p_uninstall_compat_sys_ptrace_hook(void);
-
-#endif
-
-#endif
