--- src/modules/exploit_detection/syscalls/p_sys_ptrace/p_sys_ptrace.h
@@ -1,35 +1,0 @@
-/*
- * pi3's Linux kernel Runtime Guard
- *
- * Component:
- *  - Intercept ptrace syscall
- *
- * Notes:
- *  - Enforce Exploit Detection validation
- *
- * Caveats:
- *  - None
- *
- * Timeline:
- *  - Created: 31.I.2018
- *
- * Author:
- *  - Adam 'pi3' Zabrocki (http://pi3.com.pl)
- *
- */
-
-#ifndef P_LKRG_EXPLOIT_DETECTION_SYS_PTRACE_H
-#define P_LKRG_EXPLOIT_DETECTION_SYS_PTRACE_H
-
-/* per-instance private data */
-struct p_sys_ptrace_data {
-    ktime_t entry_stamp;
-};
-
-
-int p_sys_ptrace_ret(struct kretprobe_instance *p_ri, struct pt_regs *p_regs);
-int p_sys_ptrace_entry(struct kretprobe_instance *p_ri, struct pt_regs *p_regs);
-int p_install_sys_ptrace_hook(int p_isra);
-void p_uninstall_sys_ptrace_hook(void);
-
-#endif
