+++ src/modules/print_log/p_lkrg_debug_log.c
@@ -113,9 +113,8 @@ static struct p_addr_name {
    P_LKRG_DEBUG_RULE_KPROBE(p_scm_send),
    P_LKRG_DEBUG_RULE_KPROBE(p_seccomp),
    P_LKRG_DEBUG_RULE_KPROBE(p_sys_setresgid),
-   P_LKRG_DEBUG_RULE_KPROBE(p_sys_ptrace),
+   P_LKRG_DEBUG_RULE_KPROBE(p_security_ptrace_access),
    P_LKRG_DEBUG_RULE_KPROBE(p_compat_sys_execve),
-   P_LKRG_DEBUG_RULE_KPROBE(p_compat_sys_ptrace),
    P_LKRG_DEBUG_RULE_KPROBE(p_compat_sys_add_key),
    P_LKRG_DEBUG_RULE_KPROBE(p_compat_sys_capset),
    P_LKRG_DEBUG_RULE_KPROBE(p_compat_sys_keyctl),
@@ -144,7 +143,6 @@ static struct p_addr_name {
    P_LKRG_DEBUG_RULE_KPROBE(p_call_usermodehelper),
    P_LKRG_DEBUG_RULE_KPROBE(p_sys_execveat),
    P_LKRG_DEBUG_RULE_KPROBE(p_ttwu_do_wakeup),
-   P_LKRG_DEBUG_RULE_KPROBE(p_x32_sys_ptrace),
    P_LKRG_DEBUG_RULE_KPROBE(p_x32_sys_execve),
    P_LKRG_DEBUG_RULE_KPROBE(p_x32_sys_execveat),
    P_LKRG_DEBUG_RULE_KPROBE(p_x32_sys_keyctl),
